created Api endpoint for Restaurant

+++ models/MenuItem.js
@@ -3,6 +3,6 @@ const sequelize = require('../config/database');
 
 const MenuItem = sequelize.define('MenuItem', {
     name: DataTypes.STRING,
-    price: DataTypes.FLOAT  
+    price: DataTypes.DECIMAL(10, 2) 
 });
 module.exports = MenuItem;
